Increase select timeout to be 1 second instead of 1 ms
I double someone wants to execute a syscall 1000 times per second, or check stale connections 1000 times per second, or check whether it's time to run a heartbeat 1000 times per second. Looks like there was a confusion, since it's rare to see APIs that work with microseconds instead of milliseconds. As a result I saw 500X decrease in "Waiting for CPU" metric in the profiler.