Remove SocketMode.Poll logic, don't use WinSock's select
The goal is to unify the response reading logic and let it be the same across all the platforms and usage types. Its major benefit was in better heartbeat logic that can work without thread pool and its delays, but since we can control this behavior with the previous commit, we can remove this logic entirely.

SocketMode.Poll-related logic was working only on Windows and only when connecting to Redis via non-encrypted connection. And since the increased cloud adoption such environments became extremely rare. And with unified logic it's much easier now to debug everything.