Reset DB number on EVAL in a transaction only after EXEC
Originally SE.Redis calls SELECT after each call to EVAL/EVALSHA in a transaction block to prevent having a wrong database number if a script block has changed it.
This results in "ERR EXEC without MULTI" error response in Dragonfly Redis implementation, because of an unexpected SELECT command in a transaction block.
With these changes, we'll reset the database number to the default one, but only after transaction succeeds. Users can control scripts and commands that run inside the same transaction, to prevent wrong database number to be used. For other multiplexed connections, SELECT is still here, preventing bad things to happen.